Use float instead of double

+++ gl-matrix.h
@@ -22,11 +22,11 @@ extern "C" {
                               (GL_MATRIX_MINOR_VERSION << 8) | \
                               (GL_MATRIX_MICRO_VERSION))
 
-typedef double *vec3_t;
-typedef double *vec4_t;
-typedef double *mat3_t;
-typedef double *mat4_t;
-typedef double *quat_t;
+typedef float *vec3_t;
+typedef float *vec4_t;
+typedef float *mat3_t;
+typedef float *mat4_t;
+typedef float *quat_t;
 
 /*
  * vec3_t - 3 Dimensional Vector
@@ -125,7 +125,7 @@ vec3_t vec3_negate(vec3_t vec, vec3_t dest);
  * Returns:
  * dest if not NULL, vec otherwise
  */
-vec3_t vec3_scale(vec3_t vec, double val, vec3_t dest);
+vec3_t vec3_scale(vec3_t vec, float val, vec3_t dest);
 
 /*
  * vec3_normalize
@@ -165,7 +165,7 @@ vec3_t vec3_cross (vec3_t vec, vec3_t vec2, vec3_t dest);
  * Returns:
  * Length of vec
  */
-double vec3_length(vec3_t vec);
+float vec3_length(vec3_t vec);
 
 /*
  * vec3_dot
@@ -178,7 +178,7 @@ double vec3_length(vec3_t vec);
  * Returns:
  * Dot product of vec and vec2
  */
-double vec3_dot(vec3_t vec, vec3_t vec2);
+float vec3_dot(vec3_t vec, vec3_t vec2);
 
 /*
  * vec3_direction
@@ -208,7 +208,7 @@ vec3_t vec3_direction (vec3_t vec, vec3_t vec2, vec3_t dest);
  * dest if not NULL, vec otherwise
  */
 
-vec3_t vec3_lerp(vec3_t vec, vec3_t vec2, double lerp, vec3_t dest);
+vec3_t vec3_lerp(vec3_t vec, vec3_t vec2, float lerp, vec3_t dest);
 
 /*
  * vec3_dist
@@ -221,7 +221,7 @@ vec3_t vec3_lerp(vec3_t vec, vec3_t vec2, double lerp, vec3_t dest);
  * Returns:
  * distance between vec and vec2
  */
-double vec3_dist(vec3_t vec, vec3_t vec2);
+float vec3_dist(vec3_t vec, vec3_t vec2);
 
 /*
  * vec3_unproject
@@ -393,7 +393,7 @@ mat4_t mat4_transpose(mat4_t mat, mat4_t dest);
  * Returns:
  * determinant of mat
  */
-double mat4_determinant(mat4_t mat);
+float mat4_determinant(mat4_t mat);
 
 /*
  * mat4_inverse
@@ -533,7 +533,7 @@ mat4_t mat4_scale(mat4_t mat, vec3_t vec, mat4_t dest);
  * Returns:
  * dest if not NULL, mat otherwise
  */
-mat4_t mat4_rotate(mat4_t mat, double angle, vec3_t axis, mat4_t dest);
+mat4_t mat4_rotate(mat4_t mat, float angle, vec3_t axis, mat4_t dest);
 
 /*
  * mat4_rotateX
@@ -547,7 +547,7 @@ mat4_t mat4_rotate(mat4_t mat, double angle, vec3_t axis, mat4_t dest);
  * Returns:
  * dest if not NULL, mat otherwise
  */
-mat4_t mat4_rotateX(mat4_t mat, double angle, mat4_t dest);
+mat4_t mat4_rotateX(mat4_t mat, float angle, mat4_t dest);
 
 /*
  * mat4_rotateY
@@ -561,7 +561,7 @@ mat4_t mat4_rotateX(mat4_t mat, double angle, mat4_t dest);
  * Returns:
  * dest if not NULL, mat otherwise
  */
-mat4_t mat4_rotateY(mat4_t mat, double angle, mat4_t dest);
+mat4_t mat4_rotateY(mat4_t mat, float angle, mat4_t dest);
 
 /*
  * mat4_rotateZ
@@ -575,7 +575,7 @@ mat4_t mat4_rotateY(mat4_t mat, double angle, mat4_t dest);
  * Returns:
  * dest if not NULL, mat otherwise
  */
-mat4_t mat4_rotateZ(mat4_t mat, double angle, mat4_t dest);
+mat4_t mat4_rotateZ(mat4_t mat, float angle, mat4_t dest);
 
 /*
  * mat4_frustum
@@ -590,7 +590,7 @@ mat4_t mat4_rotateZ(mat4_t mat, double angle, mat4_t dest);
  * Returns:
  * dest if not NULL, a new mat4_t otherwise
  */
-mat4_t mat4_frustum(double left, double right, double bottom, double top, double near, double far, mat4_t dest);
+mat4_t mat4_frustum(float left, float right, float bottom, float top, float near, float far, mat4_t dest);
 
 /*
  * mat4_perspective
@@ -605,7 +605,7 @@ mat4_t mat4_frustum(double left, double right, double bottom, double top, double
  * Returns:
  * dest if not NULL, a new mat4_t otherwise
  */
-mat4_t mat4_perspective(double fovy, double aspect, double near, double far, mat4_t dest);
+mat4_t mat4_perspective(float fovy, float aspect, float near, float far, mat4_t dest);
 
 /*
  * mat4_ortho
@@ -620,7 +620,7 @@ mat4_t mat4_perspective(double fovy, double aspect, double near, double far, mat
  * Returns:
  * dest if not NULL, a new mat4_t otherwise
  */
-mat4_t mat4_ortho(double left, double right, double bottom, double top, double near, double far, mat4_t dest);
+mat4_t mat4_ortho(float left, float right, float bottom, float top, float near, float far, mat4_t dest);
 
 /*
  * mat4_lookAt
@@ -721,7 +721,7 @@ quat_t quat_calculateW(quat_t quat, quat_t dest);
  *
  * @return {number} Dot product of quat and quat2
  */
-double quat_dot(quat_t quat, quat_t quat2);
+float quat_dot(quat_t quat, quat_t quat2);
 
 /*
  * quat_inverse
@@ -759,7 +759,7 @@ quat_t quat_conjugate(quat_t quat, quat_t dest);
  * Returns:
  * Length of quat
  */
-double quat_length(quat_t quat);
+float quat_length(quat_t quat);
 
 /*
  * quat_normalize
@@ -842,7 +842,7 @@ quat_t quat_toMat4(quat_t quat, mat4_t dest);
  * Returns:
  * dest if not NULL, quat otherwise
  */
-quat_t quat_slerp(quat_t quat, quat_t quat2, double slerp, quat_t dest);
+quat_t quat_slerp(quat_t quat, quat_t quat2, float slerp, quat_t dest);
 
 /*
  * quat_str
